Create editor.py
